+++ R/test_pre_trend.R
@@ -109,9 +109,10 @@ mdd_test_pre_trend_event <- function(mdd_dat, ...){
 
   ## tidy res
   ## assemble
+  name_joint <- attributes(test_joint)$heading[2:(K_before+1)]
   rbind(test_joint %>% broom::tidy() %>%
           distinct(.data$statistic, .data$p.value) %>%
-          mutate(term = paste(H_indiv, collapse = " AND ")),
+          mutate(term = paste(name_joint, collapse = " AND ")),
         test_indiv %>%
           distinct(.data$term, .data$statistic, .data$p.value)) %>%
     mutate(test = c("test_joint", rep("test_indiv", nrow(test_indiv)))) %>%